Correcting bit precision on return type

+++ IMU.cpp
@@ -44,11 +44,11 @@ void IMU::initialize(int id) {
     }
 }
 
-void IMU::getSystemStatus(uint8_t* array) {
+void IMU::getSystemStatus(uint16_t* array) {
     this->bno.getSystemStatus(&array[0], &array[1], &array[2]);
 }
 
-void IMU::getXYZ(uint8_t* array) {
+void IMU::getXYZ(uint16_t* array) {
     sensors_event_t event;
     this->bno.getEvent(&event);
     array[0] = event.orientation.x;
